requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pymysql>=1.1.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
requests>=2.31.0
python-multipart>=0.0.6
cryptography>=42.0.0
